refactor: Merge private hook with public hook
