+++ packages/insight-viewer/src/hooks/useImageLoad/types.ts
